:recycle: refactor: refactoring detail tabs
