+++ views/pool-details/accordion/accordion-icon.tsx
