+++ views/pool-details/detail-tabs/pool-details-tabs.types.ts
@@ -1,12 +1,12 @@
 import { ReactNode } from 'react';
 
-export interface PoolDetailsTabsItemProps {
+export interface DetaiTabItemProps {
   item: ReactNode;
   isSelected: boolean;
   onChange: () => void;
 }
 
-export interface PoolDetailsTabsProps {
+export interface DetaiTabsProps {
   defaultTabIndex?: number;
   items: ReadonlyArray<ReactNode>;
   onChangeTab?: (tabIndex: number) => void;
