+++ views/pool-details/detail.tsx
@@ -1,0 +1,26 @@
+import { Box } from '@interest-protocol/ui-kit';
+import { v4 } from 'uuid';
+
+import Accordion from './accordion';
+import ItemStandard from './accordion/item-standard';
+import { POOL_INFORMATION_DATA } from './pool-detail.data';
+
+const Detail = () => (
+  <Box>
+    <Accordion title={POOL_INFORMATION_DATA.title}>
+      {POOL_INFORMATION_DATA.data.map(
+        ({ label, info, additionalInfo, clipBoard }) => (
+          <ItemStandard
+            key={v4()}
+            label={label}
+            content={info}
+            isCopyClipBoard={clipBoard}
+            hasAddtionalInfo={additionalInfo}
+          />
+        )
+      )}
+    </Accordion>
+  </Box>
+);
+
+export default Detail;
