+++ views/pool-details/index.tsx
@@ -10,10 +10,10 @@ import { TOKEN_ICONS } from '@/lib';
 
 import PoolTitleBar from '../components/pool-title-bar';
 import Error from '../error';
-import PoolAdvanceDetail from './pool-advance-detail';
-import PoolDetail from './pool-detail';
+import PoolAdvanceDetail from './advance-detail';
+import PoolDetail from './detail';
+import DetailTabs from './detail-tabs';
 import { PoolDetailsProps, PoolDetailsTabOption } from './pool-details.types';
-import PoolDetailsTabs from './pool-details-tabs';
 import PoolForm from './pool-form';
 
 const PoolDetails: FC<PoolDetailsProps> = ({
@@ -65,7 +65,7 @@ const PoolDetails: FC<PoolDetailsProps> = ({
           handleOptionTab={handleOptionTab}
         />
         <Box color="onSurface" borderRadius="xs" bg="lowestContainer">
-          <PoolDetailsTabs
+          <DetailTabs
             onChangeTab={handleTabChange}
             defaultTabIndex={poolDetailsView}
             items={['Pool Detail', 'Advance Details']}
